+++ di/src/test/java/com/interface21/context/support/ConfigurationBeanScannerTest.java
@@ -1,0 +1,53 @@
+package com.interface21.context.support;
+
+import com.interface21.beans.factory.BeanFactory;
+import com.interface21.beans.factory.support.DefaultListableBeanFactory;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import samples.ExampleConfig;
+import samples.IntegrationConfig;
+import samples.JdbcSampleRepository;
+import samples.JdbcTemplate;
+
+import javax.sql.DataSource;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+class ConfigurationBeanScannerTest {
+    @Test
+    @DisplayName("클래스를 인자로 register 가 잘 동작한다")
+    public void register_simple() {
+        BeanFactory beanFactory = new DefaultListableBeanFactory();
+        ConfigurationBeanScanner cbs = new ConfigurationBeanScanner(beanFactory);
+        cbs.register(ExampleConfig.class);
+        beanFactory.initialize();
+
+        assertNotNull(beanFactory.getBean(DataSource.class));
+    }
+
+    @Test
+    public void register_classpathBeanScanner_통합() {
+        BeanFactory beanFactory = new DefaultListableBeanFactory();
+        ConfigurationBeanScanner cbs = new ConfigurationBeanScanner(beanFactory);
+        cbs.register(IntegrationConfig.class);
+
+        ClasspathBeanScanner cbds = new ClasspathBeanScanner(beanFactory);
+        cbds.doScan("samples");
+        beanFactory.initialize();
+
+        assertNotNull(beanFactory.getBean(DataSource.class));
+
+        JdbcSampleRepository sampleRepository = beanFactory.getBean(JdbcSampleRepository.class);
+        assertNotNull(sampleRepository);
+        assertNotNull(sampleRepository.getDataSource());
+
+        JdbcTemplate jdbcTemplate = beanFactory.getBean(JdbcTemplate.class);
+        assertNotNull(jdbcTemplate);
+        assertNotNull(jdbcTemplate.getDataSource());
+
+        String[] basePackages = cbds.getBasePackages();
+        assertThat(basePackages).containsExactly("samples");
+    }
+
+}
